Create Index HTML
Added a <header> with a descriptive <h1> for the website title.
Implemented a <nav> element with an unordered list (<ul>) containing navigation links (<li><a>).
I inserted an <img> element linking to CodeSpace Cat's profile picture for visual representation.
Utilized a <section> with an id of "about" to encapsulate the introductory content, including a heading (<h1>) and a paragraph describing CodeSpace Cat's passion and experience in web development.
I added another <section> with an id of "skills" to present a list of skills using an unordered list (<ul>) with list items (<li>).
Portfolio Section
I added a<section> to showcase projects under the heading "Catfolio".
Organized project details in a <table> for clarity, including columns for project name, date, description, and thumbnail images using <img> tags with corresponding alt text.
I added a <section> with an id of "contact" for users to reach out.
Integrated a <form> with action set to "#contact" and method "post" to facilitate name, email, message input fields, and a submit button (<input type="submit>).
Concluded with a <footer> containing copyright information ("© 2023 CodeSpace Cat. All rights reserved.").